+++ src/lambdaisland/glogi/console.cljs
@@ -5,10 +5,9 @@
             [goog.debug.LogBuffer :as LogBuffer]
             [goog.debug.Console :as Console]))
 
-(goog-define colorize "true")
-
-(defn colorize? []
-  (= colorize "true"))
+;; By default we do CSS colorization on non-IE browsers only. You can change
+;; this in your build with :closure-defines. Possible values: "auto" "true" "false"
+(goog-define colorize "auto")
 
 (defn log-method [level]
   (condp #(>= %2 %1) (glogi/level-value level)
@@ -18,18 +17,16 @@
     (glogi/level-value :config)  "log"
     "log"))
 
-(defn devtools-installed? []
-  (and (exists? js/devtools.core.installed_QMARK_)
-       (js/devtools.core.installed_QMARK_)))
+(defn format-raw [{:keys [level logger-name message exception]}]
+  [(str "[" logger-name "]") message])
 
-(defn format [{:keys [level logger-name message exception]}]
-  (if (devtools-installed?)
-    [(str "[" logger-name "]") message]
-    (if (colorize?)
-      (print/format level logger-name message)
-      [(str "[" logger-name "]") (pr-str message)])))
+(defn format-css [{:keys [level logger-name message exception]}]
+  (print/format level logger-name message))
 
-(defonce console-log
+(defn format-plain [{:keys [level logger-name message exception]}]
+  [(str "[" logger-name "]") (pr-str message)])
+
+(defn make-console-log [format]
   (fn [{:keys [logger-name level exception] :as record}]
     (let [method-name (log-method level)
           method      (or (gobj/get js/console method-name)
@@ -38,10 +35,47 @@
       (when exception
         (method (str "[" logger-name "]") (str exception) "\n" (.-stack exception))))))
 
+(defonce console-log-raw (make-console-log format-raw))
+(defonce console-log-css (make-console-log format-css))
+(defonce console-log-plain (make-console-log format-plain))
+
+;; backward compatibility
+(defonce format format-plain)
+(defonce console-log console-log-plain)
+
+(defn devtools-installed? []
+  (and (exists? js/devtools.core.installed_QMARK_)
+       (js/devtools.core.installed_QMARK_)))
+
+(defn- browser? []
+  (exists? js/window))
+
+(defn- ie? []
+  (and (browser?)
+       (exists? js/window.navigator)
+       (exists? js/window.navigator.userAgent)
+       ;; IE and pre-chromium EDGE don't support %c
+       (or (.includes js/window.navigator.userAgent "MSIE")
+           (.includes js/window.navigator.userAgent "Trident"))))
+
+(defn select-handler []
+  (case colorize
+    "auto"
+    (cond
+      (devtools-installed?)        console-log-raw
+      (and (browser?) (not (ie?))) console-log-css
+      :else                        console-log-plain)
+    "raw"
+    console-log-raw
+    "true"
+    console-log-css
+    "false"
+    console-log-plain))
+
 (defn install! []
   ;; Disable goog.debug.Console if it's been enabled (e.g. by Figwheel), we do
   ;; console logging now
   (when-let [instance Console/instance]
     (.setCapturing instance false))
 
-  (glogi/add-handler-once console-log))
+  (glogi/add-handler-once (select-handler)))